feat: update with latest py2hwsw, working vutb, automate actions for CI and release artifacts

+++ scripts/gen_custom_config_build.py
@@ -2,7 +2,6 @@ import os
 
 
 def gen_custom_config_build(py_params_dict):
-
     setup_dir = f"{os.path.dirname(__file__)}/pyRawWrapper/pyRawWrapper"
 
     content = f"""
@@ -43,7 +42,8 @@ IOB_CONSOLE_PYTHON_ENV ?= {setup_dir}
 endif
 """
 
-    file_path = os.path.join(py_params_dict["build_dir"], "custom_config_build.mk")
-    with open(file_path, "w") as f:
-        f.write(content)
-        f.close()
+    if py_params_dict["build_dir"] and os.path.exists(py_params_dict["build_dir"]):
+        file_path = os.path.join(py_params_dict["build_dir"], "custom_config_build.mk")
+        with open(file_path, "w") as f:
+            f.write(content)
+            f.close()
